+++ libraries/sys_helpers_nftables.rb
@@ -34,16 +34,28 @@ module Sys
         end
       end
 
-      def build_set_of_ips(ips)
-        set_of_ips = Array(ips).map { |ip| IPAddr.new(ip) }
-
+      # ips could be like '192.0.2.1', but also '$some_variable'
+      def s2ip_with_prefix(ip)
         # Only works on buster and newer. In older debian-versions
         # there is no prefix-method for IPv4-addresses.
-        addrs = set_of_ips.map { |ip| "#{ip}/#{ip.prefix}" }
-        if addrs.length == 1
-          addrs.first
+        addr = IPAddr.new(ip)
+        "#{addr}/#{addr.prefix}"
+      rescue IPAddr::InvalidAddressError
+        ip
+      end
+
+      def build_set_of_ips(ips)
+        if ips.instance_of?(String)
+          s2ip_with_prefix(ips)
         else
-          "{#{addrs.join(', ')}}"
+          ips.map! do |ip|
+            s2ip_with_prefix(ip)
+          end
+          if ips.length == 1
+            ips.first
+          else
+          "{#{ips.join(', ')}}"
+          end
         end
       end
 
@@ -81,13 +93,69 @@ module Sys
       TARGET ||= {
         accept: 'accept',
         allow: 'accept',
+        counter: 'counter',
         deny: 'drop',
         drop: 'drop',
-        log: 'log prefix "nftables:" group 0',
+        log: 'log',
         masquerade: 'masquerade',
         redirect: 'redirect',
         reject: 'reject',
       }.freeze
+
+      def nftables_command_log(rule_resource)
+        log_prefix = 'prefix '
+        log_prefix << if rule_resource.log_prefix.nil?
+                        "\"#{CHAIN[rule_resource.direction]}:\""
+                      else
+                        "\"#{rule_resource.log_prefix}\""
+                      end
+        log_group = if rule_resource.log_group.nil?
+                      nil
+                    else
+                      "group #{rule_resource.log_group} "
+                    end
+        "log #{log_prefix} #{log_group}"
+      end
+
+      def nftables_command_redirect(rule_resource)
+        if rule_resource.redirect_port.nil?
+          raise 'Specify redirect_port when using :redirect as commmand'
+        end
+
+        "redirect to #{rule_resource.redirect_port} "
+      end
+
+      def nftables_commands(rule_resource)
+        nftables_rule = ''
+        commands = Array(rule_resource.command).map do |command|
+          begin
+            TARGET.fetch(command.to_sym)
+          rescue KeyError
+            raise "Invalid command: #{command.inspect}. Use one of #{TARGET.keys}"
+          end
+        end
+
+        commands.sort!.uniq!
+        # Add non-terminal statements first
+        if commands.include?('log')
+          nftables_rule << nftables_command_log(rule_resource)
+          commands -= ['log']
+        end
+        if commands.include?('counter')
+          nftables_rule << 'counter '
+          commands -= ['counter']
+        end
+        raise "Only one terminal statement is possible, provided were #{commands.join(',')}" if commands.length > 1
+        commands.each do |cmd|
+          nftables_rule << case cmd
+                           when 'redirect'
+                             nftables_command_redirect(rule_resource)
+                           else
+                             "#{cmd} "
+                           end
+        end
+        nftables_rule
+      end
 
       def build_nftables_rule(rule_resource)
         return rule_resource.raw.strip if rule_resource.raw
@@ -128,13 +196,14 @@ module Sys
           nftables_rule << "#{rule_resource.protocol} dport #{port_to_s(rule_resource.dport)} " if rule_resource.dport
         when :esp, :ah
           nftables_rule << "#{ip_family} #{ip_family == :ip6 ? 'nexthdr' : 'protocol'} #{rule_resource.protocol} "
-
-          # nothing to do default :ipv6, :none
+        when :none
+          nil # do nothing
+        else
+          nftables_rule << "#{ip_family} protocol #{rule_resource.protocol} "
         end
 
         nftables_rule << "ct state #{Array(rule_resource.stateful).join(',').downcase} " if rule_resource.stateful
-        nftables_rule << "#{TARGET[rule_resource.command.to_sym]} "
-        nftables_rule << " to #{rule_resource.redirect_port} " if rule_resource.command == :redirect
+        nftables_rule << nftables_commands(rule_resource)
         nftables_rule << "comment \"#{rule_resource.description}\" " if rule_resource.include_comment
         nftables_rule.strip!
         nftables_rule
@@ -162,7 +231,7 @@ module Sys
 
       def ensure_default_rules_exist(new_resource)
         input = new_resource.rules || {}
-        input.merge!(default_ruleset(new_resource))
+        new_resource.rules = input.merge!(default_ruleset(new_resource))
       end
     end
   end