+++ test/fixtures/cookbooks/nftables-test/recipes/default.rb
@@ -25,6 +25,18 @@ nftables_rule 'established' do
   command :allow
 end
 
+nftables_rule 'named set' do
+  position 39
+  raw "add set inet filter networks { type ipv4_addr; flags constant, interval; elements={ 192.0.2.0/24 } }"
+end
+
+nftables_rule 'variable' do
+  raw 'define quad9 = 9.9.9.9'
+end
+
+nftables_rule 'test named set' do
+  source '@networks'
+end
 
 nftables_rule 'ssh22' do
   dport 22
@@ -168,4 +180,21 @@ nftables_rule 'redirect' do
   dport 5555
   redirect_port 6666
   command :redirect
+end
+
+nftables_rule 'log_without_prefix' do
+  dport 1
+  command :log
+end
+
+nftables_rule 'log_with_prefix' do
+  dport 1
+  log_prefix 'nflog by chef:'
+  log_group 1
+  command :log
+end
+
+nftables_rule 'multiple commands' do
+  dport 2
+  command [:counter, :log, :accept]
 end